chore: move FactionTabs to shared/components, break archives↔dashboard cycle
FactionTabs is a generic UI primitive (faction tab selector) used by
both dashboard and archives. Moving to shared/ eliminates the last
bidirectional dependency between these features.
