+++ src/features/dashboard/DashboardClient.jsx
@@ -4,7 +4,7 @@ import NotificationToggle from '@/features/notifications/NotificationToggle';
 import LastUpdated from '@/shared/components/LastUpdated';
 import EventCard, { computeFrontier } from '@/features/galaxy/EventCard';
 import DefeatedCard from '@/features/galaxy/DefeatedCard';
-import FactionTabs from '@/features/dashboard/FactionTabs';
+import FactionTabs from '@/shared/components/FactionTabs';
 import RegionsViewToggle from '@/features/dashboard/RegionsViewToggle';
 import StatGrid from '@/features/stats/StatGrid';
 import { useLiveDataContext } from '@/shared/providers/LiveDataContext.mjs';
