+++ src/shared/components/FactionTabs.jsx
